+++ maia/pytree/sids/node_inspect.py
@@ -1,6 +1,7 @@
 import numpy as np
 import itertools
 
+from maia.pytree.typing import *
 
 from maia.pytree.compare import check_is_label, check_in_labels
 from maia.pytree         import node as N
@@ -14,22 +15,22 @@ from .utils import for_all_methods
 class Zone:
 
   @staticmethod
-  def IndexDimension(zone_node):
+  def IndexDimension(zone_node:CGNSTree) -> int:
     z_sizes = N.get_value(zone_node)
     return (z_sizes[:,0]).size
 
   @staticmethod
-  def VertexSize(zone_node):
+  def VertexSize(zone_node:CGNSTree) -> Union[int, np.ndarray]:
     sizes = N.get_value(zone_node)[:,0]
     return sizes[0] if Zone.Type(zone_node) == 'Unstructured' else sizes
 
   @staticmethod
-  def CellSize(zone_node):
+  def CellSize(zone_node:CGNSTree) -> Union[int, np.ndarray]:
     sizes = N.get_value(zone_node)[:,1]
     return sizes[0] if Zone.Type(zone_node) == 'Unstructured' else sizes
 
   @staticmethod
-  def FaceSize(zone_node):
+  def FaceSize(zone_node:CGNSTree) -> Union[int, np.ndarray]:
 
     def compute_nface_per_direction(d, dim, vtx_size, cell_size):
       n_face_per_direction = vtx_size[d%dim]
@@ -55,55 +56,55 @@ class Zone:
     return n_face
 
   @staticmethod
-  def NGonNode(zone_node):
+  def NGonNode(zone_node:CGNSTree) -> CGNSTree:
     predicate = lambda n: N.get_label(n) == "Elements_t" and Element.CGNSName(n) == 'NGON_n'
     ngons = W.get_children_from_predicate(zone_node, predicate)
     return utils.expects_one(ngons, ("NGon node", f"zone {N.get_name(zone_node)}"))
 
   @staticmethod
-  def NFaceNode(zone_node):
+  def NFaceNode(zone_node:CGNSTree) -> CGNSTree:
     predicate = lambda n: N.get_label(n) == "Elements_t" and Element.CGNSName(n) == 'NFACE_n'
     nfaces = W.get_children_from_predicate(zone_node, predicate)
     return utils.expects_one(nfaces, ("NFace node", f"zone {N.get_name(zone_node)}"))
 
   @staticmethod
-  def VertexBoundarySize(zone_node):
+  def VertexBoundarySize(zone_node:CGNSTree) -> Union[int, np.ndarray]:
     sizes = N.get_value(zone_node)[:,2]
     return sizes[0] if Zone.Type(zone_node) == 'Unstructured' else sizes
 
   @staticmethod
-  def Type(zone_node):
+  def Type(zone_node:CGNSTree) -> str:
     zone_type_node = W.get_child_from_label(zone_node, "ZoneType_t")
     return N.get_value(zone_type_node)
 
   @staticmethod
-  def n_vtx(zone_node):
+  def n_vtx(zone_node:CGNSTree) -> int:
     return np.prod(Zone.VertexSize(zone_node))
 
   @staticmethod
-  def n_cell(zone_node):
+  def n_cell(zone_node:CGNSTree) -> int:
     return np.prod(Zone.CellSize(zone_node))
 
   @staticmethod
-  def n_face(zone_node):
+  def n_face(zone_node:CGNSTree) -> int:
     return np.sum(Zone.FaceSize(zone_node))
 
   @staticmethod
-  def n_vtx_bnd(zone_node):
+  def n_vtx_bnd(zone_node:CGNSTree) -> int:
     return np.prod(Zone.VertexBoundarySize(zone_node))
 
   @staticmethod
-  def has_ngon_elements(zone_node):
+  def has_ngon_elements(zone_node: CGNSTree) -> bool:
     predicate = lambda n: N.get_label(n) == "Elements_t" and Element.CGNSName(n) == 'NGON_n'
     return W.get_child_from_predicate(zone_node, predicate) is not None
 
   @staticmethod
-  def has_nface_elements(zone_node):
+  def has_nface_elements(zone_node:CGNSTree) -> bool:
     predicate = lambda n: N.get_label(n) == "Elements_t" and Element.CGNSName(n) == 'NFACE_n'
     return W.get_child_from_predicate(zone_node, predicate) is not None
 
   @staticmethod
-  def coordinates(zone_node, name=None):
+  def coordinates(zone_node:CGNSTree, name:str=None) -> Tuple[Optional[np.ndarray], Optional[np.ndarray], Optional[np.ndarray]]:
     grid_coord_node = W.get_child_from_label(zone_node, "GridCoordinates_t") if name is None \
         else W.get_child_from_name_and_label(zone_node, name, "GridCoordinates_t")
     if grid_coord_node is None:
@@ -119,7 +120,7 @@ class Zone:
     return x, y, z
 
   @staticmethod
-  def ngon_connectivity(zone_node):
+  def ngon_connectivity(zone_node:CGNSTree) -> List[np.ndarray]:
     ngon_node = Zone.NGonNode(zone_node)
     face_vtx_idx = N.get_value(W.get_child_from_name(ngon_node, "ElementStartOffset"))
     face_vtx     = N.get_value(W.get_child_from_name(ngon_node, "ElementConnectivity"))
@@ -127,14 +128,14 @@ class Zone:
     return face_vtx_idx, face_vtx, ngon_pe
 
   @staticmethod
-  def get_range_of_ngon(zone):
+  def get_range_of_ngon(zone:CGNSTree) -> np.ndarray:
     """
     Return the ElementRange array of the NGON elements
     """
     return Element.Range(Zone.NGonNode(zone))
 
   @staticmethod
-  def get_ordered_elements(zone):
+  def get_ordered_elements(zone:CGNSTree) -> List[CGNSTree]:
     """
     Return the elements nodes in increasing order wrt ElementRange
     """
@@ -142,7 +143,7 @@ class Zone:
                   key = lambda item : Element.Range(item)[0])
 
   @staticmethod
-  def get_ordered_elements_per_dim(zone):
+  def get_ordered_elements_per_dim(zone:CGNSTree) -> List[List[CGNSTree]]:
     """
     Return a list of size 4 containing Element nodes belonging to each dimension.
     In addition, Element are sorted according to their ElementRange withing each dimension.
@@ -151,7 +152,7 @@ class Zone:
     return utils.bucket_split(Zone.get_ordered_elements(zone), lambda e: Element.Dimension(e), size=4)
 
   @staticmethod
-  def get_elt_range_per_dim(zone):
+  def get_elt_range_per_dim(zone:CGNSTree) -> List[List[int]]:
     """
     Return a list of size 4 containing min & max element id of each dimension
     This function is relevant only if Element of same dimension have consecutive
@@ -174,7 +175,7 @@ class Zone:
     return range_by_dim
 
   @staticmethod
-  def elt_ordering_by_dim(zone):
+  def elt_ordering_by_dim(zone:CGNSTree):
     """Returns 1 if lower dimension elements have lower element range, -1 if
     lower dim. elements have higher element range, and 0 if order can not be determined"""
     status = 0
@@ -192,27 +193,27 @@ class Zone:
 class Element:
 
   @staticmethod
-  def Type(elt_node):
+  def Type(elt_node:CGNSTree) -> int:
     return int(elt_node[1][0])
 
   @staticmethod
-  def CGNSName(elt_node):
+  def CGNSName(elt_node:CGNSTree) -> str:
     return EU.element_name(Element.Type(elt_node))
 
   @staticmethod
-  def Dimension(elt_node):
+  def Dimension(elt_node:CGNSTree) -> int:
     return EU.element_dim(Element.Type(elt_node))
 
   @staticmethod
-  def NVtx(elt_node):
+  def NVtx(elt_node:CGNSTree) -> int:
      return EU.element_number_of_nodes(Element.Type(elt_node))
 
   @staticmethod
-  def Range(elt_node):
+  def Range(elt_node:CGNSTree) -> np.ndarray:
     return W.get_child_from_name(elt_node,"ElementRange")[1]
 
   @staticmethod
-  def Size(elt_node):
+  def Size(elt_node:CGNSTree) -> int:
     er = Element.Range(elt_node)
     return er[1] - er[0] + 1
 
@@ -222,7 +223,7 @@ class Element:
 class GridConnectivity:
 
   @staticmethod
-  def Type(gc):
+  def Type(gc:CGNSTree) -> str:
     if N.get_label(gc) == 'GridConnectivity1to1_t':
       return 'Abutting1to1'
     elif N.get_label(gc) == 'GridConnectivity_t':
@@ -230,15 +231,15 @@ class GridConnectivity:
       return N.get_value(gc_type_n) if gc_type_n is not None else 'Overset'
 
   @staticmethod
-  def is1to1(gc):
+  def is1to1(gc:CGNSTree) -> bool:
     return GridConnectivity.Type(gc) == 'Abutting1to1'
 
   @staticmethod
-  def isperiodic(gc):
+  def isperiodic(gc:CGNSTree) -> bool:
     return W.get_node_from_label(gc, 'Periodic_t', depth=[2,2]) is not None
 
   @staticmethod
-  def periodic_values(gc):
+  def periodic_values(gc:CGNSTree) -> Union[List[None], List[np.ndarray]]:
     """ Return periodic data if existing (RotationCenter, RotationAngle, Translation) """
     perio_node = W.get_node_from_label(gc, "Periodic_t", depth=[2,2])
     if perio_node is None:
@@ -262,17 +263,17 @@ class Subset:
   A subset is a node having a PointList or a PointRange
   """
   
-  def getPatch(subset):
+  def getPatch(subset:CGNSTree) -> CGNSTree:
     pl = W.get_child_from_name(subset, 'PointList')
     pr = W.get_child_from_name(subset, 'PointRange')
     assert (pl is None) ^ (pr is None)
     return pl if pl is not None else pr
 
-  def n_elem(subset):
+  def n_elem(subset:CGNSTree) -> int:
     patch = Subset.getPatch(subset)
     return PointList.n_elem(patch) if N.get_label(patch) == 'IndexArray_t' else PointRange.n_elem(patch)
 
-  def GridLocation(subset):
+  def GridLocation(subset:CGNSTree) -> str:
     grid_loc_n = W.get_child_from_label(subset, 'GridLocation_t')
     return N.get_value(grid_loc_n) if grid_loc_n else 'Vertex'
 
@@ -281,7 +282,7 @@ class Subset:
 class PointRange:
 
   @staticmethod
-  def SizePerIndex(point_range_node):
+  def SizePerIndex(point_range_node:CGNSTree) -> np.ndarray:
     """
     Allow point_range to be inverted (PR[:,1] < PR[:,0]) as it can occurs in struct GCs
     """
@@ -289,7 +290,7 @@ class PointRange:
     return np.abs(pr_values[:,1] - pr_values[:,0]) + 1
 
   @staticmethod
-  def n_elem(point_range_node):
+  def n_elem(point_range_node:CGNSTree) -> int:
     return PointRange.SizePerIndex(point_range_node).prod()
 
 
@@ -298,6 +299,6 @@ class PointRange:
 class PointList:
 
   @staticmethod
-  def n_elem(point_list_node):
+  def n_elem(point_list_node:CGNSTree) -> int:
     return N.get_value(point_list_node).shape[1]