+++ maia/utils/ndarray/np_utils.py
@@ -209,15 +209,53 @@ def transform_cart_matrix(vectors, translation=np.zeros(3), rotation_center=np.z
   rotation_matz = np.array([[np.cos(gamma), -np.sin(gamma), 0], [np.sin(gamma), np.cos(gamma), 0], [0, 0, 1]])
   rotation_mat  = np.dot(rotation_matx, np.dot(rotation_maty, rotation_matz))
   rotated_vectors = ((np.dot(rotation_mat, vectors-rotation_center)+rotation_center).T + translation).T
-  return (rotated_vectors.astype(vectors.dtype,copy=False))
+  return rotated_vectors.astype(vectors.dtype,copy=False)
 
 def transform_cart_vectors(vx, vy, vz, translation=np.zeros(3), rotation_center=np.zeros(3), rotation_angle=np.zeros(3)):
   """
   Apply the defined cartesian transformation on separated components of vectors and return a tuple with each of the modified components of the vectors
   """
-  vectors = np.array([vx,vy,vz], order='F')
+  assert vx.shape == vy.shape == vz.shape
+  if vx.ndim == 1:
+    vectors = np.array([vx,vy,vz], order='F')
+  else: #Manage structured blocks
+    vectors = np.array([vx.flatten('F'), vy.flatten('F'), vz.flatten('F')], order='F')
+  
   modified_components = transform_cart_matrix(vectors, translation, rotation_center, rotation_angle)
-  return (modified_components[0], modified_components[1], modified_components[2])
+
+  if vx.ndim == 1:
+    return (modified_components[0], modified_components[1], modified_components[2])
+  else:
+    return (modified_components[0].reshape(vx.shape, order='F'),
+            modified_components[1].reshape(vy.shape, order='F'),
+            modified_components[2].reshape(vz.shape, order='F'))
+
+def transform_cart_matrix_2d(vectors, translation=np.zeros(2), rotation_center=np.zeros(2), rotation_angle=0.):
+  """
+  Apply the defined cartesian transformation on 2D concatenated components of vectors described by :
+  [vx1 vx2 ... vxN]
+  [vy1 vy2 ... vyN]
+  and return the modified components of the vectors in the same format
+  """
+  theta = rotation_angle
+  rotation_mat = np.array([[np.cos(theta), -np.sin(theta)], [np.sin(theta), np.cos(theta)]])
+  _rotation_center = np.array(rotation_center).reshape((-1,1))
+  _translation = np.array(translation).reshape((-1,1))
+  rotated_vectors = (np.dot(rotation_mat, vectors-_rotation_center)+_rotation_center) + _translation
+  return rotated_vectors.astype(vectors.dtype,copy=False)
+
+
+def transform_cart_vectors_2d(vx, vy, translation=np.zeros(2), rotation_center=np.zeros(2), rotation_angle=0.):
+  assert vx.shape == vy.shape
+  if vx.ndim == 1:
+    vectors = np.array([vx,vy], order='F')
+  else: #Manage structured blocks
+    vectors = np.array([vx.flatten('F'), vy.flatten('F')], order='F')
+  modified_components = transform_cart_matrix_2d(vectors, translation, rotation_center, rotation_angle)
+  if vx.ndim == 1:
+    return (modified_components[0], modified_components[1])
+  else:
+    return (modified_components[0].reshape(vx.shape, order='F'), modified_components[1].reshape(vy.shape, order='F'))
 
 
 def safe_int_cast(array, dtype):