+++ maia/utils/ndarray/subset_sum.pybind.hpp
@@ -1,4 +1,5 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/numpy.h>
+#include "pdm.h"
 
-auto search_subset_match(pybind11::array_t<int> sorted_np, int target, int max_it) -> pybind11::list;
+auto search_subset_match(pybind11::array_t<PDM_g_num_t> sorted_np, int target, int max_it) -> pybind11::list;